frontend/requirements.txt
dash==2.14.1
dash-bootstrap-components==1.5.0
plotly==5.17.0
pandas==2.1.1
paramiko==3.4.0
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
httpx==0.24.1
kafka-python==2.0.2
confluent-kafka==2.3.0
pyspark==3.0.3
py4j==0.10.9
